Hi block utils functions file
